Abstract the gameModeType into a seperate module refactore the playGame function
